ColumnConfig isn't sql specific
