+++ corehq/apps/userreports/reports/specs.py
@@ -28,8 +28,9 @@ from sqlagg.columns import (
     YearColumn,
 )
 from corehq.apps.reports.sqlreport import DatabaseColumn, AggregateColumn
+from corehq.apps.userreports.columns import ColumnConfig
 from corehq.apps.userreports.specs import TypeProperty
-from corehq.apps.userreports.sql import get_expanded_column_config, ColumnConfig
+from corehq.apps.userreports.sql import get_expanded_column_config
 from corehq.apps.userreports.transforms.factory import TransformFactory
 from corehq.apps.userreports.util import localize
 from dimagi.utils.decorators.memoized import memoized
